feat(#3338): ALB, CLB and NLB integration tests (#3711)
* feat(#3338): Add basic test for aws ALB integration

* feat(#3338): Add remove_from_bucket test for aws ALB integration

* feat(#3338): Add schema for ALB table

* feat(#3338): Add only_logs_after tests for aws ALB integration

* feat(#3338): Add path tests for aws ALB integration

* feat(#3338): Add regions tests for aws ALB integration

* feat(#3338): Add regions discard_regex for aws ALB integration

* feat(#3338): Add regions only_logs_after tier 1 for aws ALB integration

* feat(#3338): Improve upload_and_delete_file_to_s3 fixture

* feat(#3338): Skip remove_from_bucket test case for ALB

* feat(#3338): Add basic test for aws CLB integration

* feat(#3338): Add remove_from_bucket test for aws CLB integration

* feat(#3338): Add only_logs_after tests for aws CLB integration

* feat(#3338): Add path tests for aws CLB integration

* feat(#3338): Add regions only_logs_after tier 1 for aws CLB integration

* feat(#3338): Add regions discard_regex for aws CLB integration

* feat(#3338): Add only_logs_after tier 1 for aws CLB integration

* feat(#3338): Skip remove_from_bucket test case for CLB

* feat(#3338): Add basic test for aws NLB integration

* feat(#3338): Add remove_from_bucket test for aws NLB integration

* feat(#3338): Add only_logs_after tests for aws NLB integration

* feat(#3338): Add path tests for aws NLB integration

* feat(#3338): Add regions tests for aws NLB integration

* feat(#3338): Add only_logs_after tier 1 for aws NLB integration

* style(#3335): Fix linter issues

* feat(#3338): Move skip function to a fixture

* feat(#3338): Use join in path builds

* feat(#3338): Improve docstrings

* feat(#3338): Increase timeout to avoid random failures

* refactor(#3338): Use utcnow() instead of now() and typos

* refactor(#3338): Apply changes from CR

+++ tests/integration/test_aws/conftest.py
@@ -3,7 +3,7 @@ from typing import Generator
 import pytest
 from wazuh_testing import UDP, logger
 from wazuh_testing.modules.aws.db_utils import delete_s3_db
-from wazuh_testing.modules.aws.s3_utils import delete_file, upload_file
+from wazuh_testing.modules.aws.s3_utils import delete_file, upload_file, file_exists
 from wazuh_testing.tools import ANALYSISD_QUEUE_SOCKET_PATH, LOG_FILE_PATH
 from wazuh_testing.tools.file import bind_unix_socket
 from wazuh_testing.tools.monitoring import FileMonitor, ManInTheMiddle, QueueMonitor
@@ -33,6 +33,12 @@ def analysisd_monitor() -> Generator:
 
     mitm.shutdown()
     control_service('start', daemon='wazuh-analysisd')
+
+
+@pytest.fixture(scope='function')
+def mark_cases_as_skipped(metadata: dict) -> None:
+    if metadata['name'] in ['alb_remove_from_bucket', 'clb_remove_from_bucket', 'nlb_remove_from_bucket']:
+        pytest.skip(reason='ALB, CLB and NLB integrations are removing older logs from other region')
 
 
 # S3 fixtures
@@ -66,8 +72,9 @@ def upload_and_delete_file_to_s3(metadata: dict):
 
     yield
 
-    delete_file(filename=filename, bucket_name=bucket_name)
-    logger.debug('Deleted file: %s from bucket %s', filename, bucket_name)
+    if file_exists(filename=filename, bucket_name=bucket_name):
+        delete_file(filename=filename, bucket_name=bucket_name)
+        logger.debug('Deleted file: %s from bucket %s', filename, bucket_name)
 
 
 @pytest.fixture(scope='function')
